REF: Move qt_kicker into utils.
- For now, `from qt_kicker import install_qt_kicker` raises
  a warning.
